# docs/swagger.yaml
basePath: /
definitions:
...
  dto.BaseResponse:
    properties:
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.NotificationListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.Notification'
        type: array
      message:
        type: string
      page:
        type: integer
      success:
        example: true
        type: boolean
      total:
        type: integer
    type: object
  dto.NotificationSendSummary:
    properties:
      failed_count:
        type: integer
      sent_count:
        type: integer
      total_processed:
        type: integer
    type: object
  dto.NotificationStatusSummary:
    properties:
      delivered_count:
        type: integer
      failed_count:
        type: integer
      still_processing:
        type: integer
      total_checked:
        type: integer
    type: object
...
  entity.DeliveryStatus:
    enum:
    - QUEUED
    - SENT
    - DELIVERED
    - FAILED
    - RESEND
    - CANCELLED
    type: string
    x-enum-varnames:
    - DeliveryQueued
    - DeliverySent
    - DeliveryDelivered
    - DeliveryFailed
    - DeliveryResend
    - DeliveryCancelled
...
  entity.Notification:
    properties:
      content:
        type: string
      created_at:
        type: string
      delivery_status:
        $ref: '#/definitions/entity.DeliveryStatus'
      failure_reason:
        type: string
      id:
        type: string
      notification_type:
        $ref: '#/definitions/entity.NotificationType'
      phone_number:
        type: string
      provider_message_id:
        type: string
      referral_id:
        type: string
      retry_count:
        type: integer
      sent_at:
        type: string
    type: object
  entity.NotificationType:
    enum:
    - ACCEPTANCE
    - SCHEDULING
    - REMINDER
    - RESCHEDULE
    type: string
    x-enum-varnames:
    - NotifyAcceptance
    - NotifyScheduling
    - NotifyReminder
    - NotifyReschedule
...
info:
  contact: {}
  description: |-
    # Hospital Referral Hub API
    A national‑scale hospital referral management platform that digitises the entire patient‑transfer workflow, from initial doctor referral to final clinical outcome. All actions are governed by strict role‑based access controls and clinical governance rules.

    ---
    ## Referral Lifecycle
    DRAFT → SUBMITTED → (Liaison) UNDER_LIAISON_REVIEW → FORWARDED
    ↘ REJECTED_BY_LIAISON
    FORWARDED → (Specialist) UNDER_SPECIALIST_REVIEW → ACCEPTED / REJECTED_BY_SPECIALIST
    ACCEPTED → SCHEDULED → ASSIGNED → COMPLETED
    ↘ MISSED / RESCHEDULED / DECEASED

    ---
    ## Visibility Rules

    | Status | Visible To |
    |----------------------|-----------|
    | DRAFT / NEED_REVISION | Referring Doctor only |
    | SUBMITTED … FORWARDED | Liaison of the sender hospital |
    | FORWARDED … COMPLETED | Specialists of the target hospital |
    | ACCEPTED … RESCHEDULED| Receptionists of the target hospital |
    | All statuses | System Admins (global); MoH Analysts (aggregated dashboards, no raw clinical data) |

    ---
    ## Critical Business Rules

    - **ML Triage Gate**: A referral cannot be accepted without a severity score (set manually via `POST /specialist/referrals/{id}/triage-severity`).
...
    - **Walk‑in Restriction**: Walk‑ins can only be registered for referrals in ACCEPTED or SCHEDULED status.
...

    For detailed per‑endpoint rules, see the individual endpoint descriptions below.
...
paths:
...
  /api/v1/internal/jobs/extend-daily-schedule:
    post:
      description: |-
        Triggers the expansion of the rolling capacity window. Typically called by a nightly cron job.
        **Roles:** SYSTEM_SUPER_ADMIN, HOSPITAL_ADMIN, DEPT_HEAD
        **Prerequisites:** Authenticated administrative session.
        **Common Errors:**
        - 401 Unauthorized
        - 500 Internal Server Error
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - BearerAuth: []
      summary: Extend Daily Schedule Window
      tags:
      - Internal/Jobs
  /api/v1/internal/jobs/run-scheduler-cycle:
    post:
      description: |-
        Run one sharded scheduler cycle (processes one department).
        **Roles:** SYSTEM_SUPER_ADMIN, HOSPITAL_ADMIN, DEPT_HEAD
        **State Transition:** Moves WAITING referrals to SCHEDULED if capacity exists.
        **Common Errors:**
        - 401 Unauthorized
        - 500 Internal Server Error
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Run Automated Batch Scheduling Cycle
      tags:
      - Internal/Jobs
  /api/v1/internal/jobs/send-reminders:
    post:
      description: |-
        Finds appointments for tomorrow and queues SMS reminders for patients.
        **Roles:** SYSTEM_SUPER_ADMIN, HOSPITAL_ADMIN, DEPT_HEAD
        **Prerequisites:** Authenticated administrative session.
        **Common Errors:**
        - 401 Unauthorized
        - 500 Internal Server Error
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Batch Queue Appointment Reminders
      tags:
      - Internal/Jobs
  /api/v1/internal/jobs/update-waiting-weights:
    post:
      description: |-
        Increments waiting_hours_weight for all WAITING referrals.
        **Roles:** SYSTEM_SUPER_ADMIN, HOSPITAL_ADMIN, DEPT_HEAD
        **Constraints:** Idempotent (runs once per calendar day).
        **Common Errors:**
        - 401 Unauthorized
        - 500 Internal Server Error
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - BearerAuth: []
      summary: Manual Trigger for Waiting Weight Updates
      tags:
      - Internal/Jobs
  /api/v1/internal/notifications:
    get:
      description: |-
        Retrieve a paginated list of notifications with various filters.
        **Roles:** SYSTEM_SUPER_ADMIN, HOSPITAL_ADMIN, DEPT_HEAD, RECEPTIONIST
        **Scoping:**
        - SYSTEM_SUPER_ADMIN: Global access, can filter by any hospital/department.
        - HOSPITAL_ADMIN: Scoped to their hospital.
        - DEPT_HEAD / RECEPTIONIST: Scoped to their hospital and department.
        **Common Errors:**
        - 401 Unauthorized
        - 403 Forbidden
        - 500 Internal Server Error
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by Referral ID
        in: query
        name: referral_id
        type: string
      - description: Filter by Hospital ID (Ignored if not Super Admin)
        in: query
        name: hospital_id
        type: string
      - description: Filter by Department ID (Ignored if not Super Admin/Hospital
          Admin)
        in: query
        name: department_id
        type: string
      - description: Filter by Notification Type
        in: query
        name: notif_type
        type: string
      - description: Filter by Delivery Status
        in: query
        name: delivery_status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NotificationListResponse'
      security:
      - BearerAuth: []
      summary: List Notifications
      tags:
      - Internal/Notifications
  /api/v1/internal/notifications/{id}/resend:
    post:
      description: |-
        Resends a failed or queued notification.
        **Roles:** SYSTEM_SUPER_ADMIN, HOSPITAL_ADMIN, DEPT_HEAD
        **Common Errors:**
        - 400 invalid notification ID
        - 401 Unauthorized
        - 500 Internal Server Error
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - BearerAuth: []
      summary: Manual Resend Notification
      tags:
      - Internal/Notifications
  /api/v1/internal/notifications/send:
    post:
      description: |-
        Processes up to 50 queued SMS notifications. Optionally filter by hospital/department.
        **Roles:** SYSTEM_SUPER_ADMIN, HOSPITAL_ADMIN, DEPT_HEAD
        **Prerequisites:** Authenticated administrative session.
        **Common Errors:**
        - 401 Unauthorized
        - 500 Internal Server Error
      parameters:
      - description: Filter by Hospital ID
        in: query
        name: hospital_id
        type: string
      - description: Filter by Department ID
        in: query
        name: department_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NotificationSendSummary'
      security:
      - BearerAuth: []
      summary: Manual Trigger Notification Send
      tags:
      - Internal/Notifications
  /api/v1/internal/notifications/update-status:
    post:
      description: |-
        Sync delivery status of SENT notifications with the SMS provider (AfroMessage).
        **Roles:** SYSTEM_SUPER_ADMIN, HOSPITAL_ADMIN, DEPT_HEAD
        **Common Errors:**
        - 401 Unauthorized
        - 500 Internal Server Error
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NotificationStatusSummary'
      security:
      - BearerAuth: []
      summary: Manual Sync Delivery Status
      tags:
      - Internal/Notifications
  /api/v1/internal/notifications/webhook:
    get:
      description: |-
        Endpoint for AfroMessage to push status updates.
        **Roles:** SMS_PROVIDER (Public)
        **Common Errors:**
        - 500 Internal Server Error
      parameters:
      - description: Provider Message ID
        in: query
        name: message_id
        required: true
        type: string
      - description: Delivery Status
        in: query
        name: status
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: SMS Webhook
      tags:
      - Internal/Notifications
...
securityDefinitions:
  BearerAuth:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"